Time: 152 ms (73.99%), Space: 75.5 MB (86.16%) - LeetHub

+++ 2537-count-the-number-of-good-subarrays/2537-count-the-number-of-good-subarrays.cpp
@@ -6,7 +6,7 @@ class Solution {
 public:
     long long countGood(vector<int>& nums, int k) {
         long long res = 0, n = nums.size(), pairs = 0;
-        unordered_map<int, long long> mp;
+        unordered_map<int, int> mp;
         for (int i = 0, j = 0; i < n; i++)
         {
             pairs += mp[nums[i]]++;
